start, refactor, move webhook related code into own module
